Added restriction for removing budget plan

+++ src/FamilyBudget.Application/Features/Finances/Budget/Commands/Handlers/DeleteBudgetPlanHandler.cs
@@ -1,6 +1,7 @@
 using FamilyBudget.Common.Abstractions.Communication;
 using FamilyBudget.Common.Results;
 using FamilyBudget.Domain.Entities.Budget;
+using FamilyBudget.Domain.Interfaces.Auth;
 using FamilyBudget.Domain.Interfaces.Repositories;
 using FamilyBudget.Domain.Interfaces.Repositories.Base;
 
@@ -9,11 +10,16 @@ namespace FamilyBudget.Application.Features.Finances.Budget.Commands.Handlers;
 public sealed class DeleteBudgetPlanHandler : ICommandHandler<DeleteBudgetPlan>
 {
     private readonly IBudgetPlanRepository _budgetPlanRepository;
+    private readonly IUserContext _userContext;
     private readonly IUnitOfWork _unitOfWork;
 
-    public DeleteBudgetPlanHandler(IBudgetPlanRepository budgetPlanRepository, IUnitOfWork unitOfWork)
+    public DeleteBudgetPlanHandler(
+        IBudgetPlanRepository budgetPlanRepository,
+        IUserContext userContext,
+        IUnitOfWork unitOfWork)
     {
         _budgetPlanRepository = budgetPlanRepository;
+        _userContext = userContext;
         _unitOfWork = unitOfWork;
     }
 
@@ -21,6 +27,8 @@ public sealed class DeleteBudgetPlanHandler : ICommandHandler<DeleteBudgetPlan>
     {
         var budgetPlan = await _budgetPlanRepository.GetByIdAsync(request.BudgetPlanId);
         if (budgetPlan is null) return Result.NotFound(nameof(BudgetPlan), request.BudgetPlanId);
+
+        if (budgetPlan.Creator.ExternalId != _userContext.UserId) return Result.Unauthorized();
 
         _budgetPlanRepository.Remove(budgetPlan);
         var result = await _unitOfWork.CommitAsync();
